grpc-voice/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.0
grpcio-tools==1.59.0
grpcio-reflection==1.59.0
openai-whisper==20231117
pydub==0.25.1
python-dotenv==1.0.0
pika==1.3.2
aio-pika==9.3.1
httpx==0.25.1
pydantic-settings==2.1.0
pydantic==2.5.0
